Refactor repository structure and coach module
- Add MIT LICENSE
- Rename Data/ to data/ for convention consistency
- Split coach/agent.py into modular package (constants, state, inference, detection, diagnostics)
- Consolidate prompts/ into coach/prompts.py
- Move main.py to scripts/cli.py
- Add pyproject.toml for project config
- Fix Qdrant vector config (text-dense)
- Fix citation logic for educational response mode
- Gitignore local-only docs (AI_GUIDE, SECURITY_HARDENING)
