+++ coach/detection/__init__.py
@@ -1,0 +1,5 @@
+"""Detection module for user intent and emotional state detection.
+
+Import directly from detection.detectors to avoid circular imports:
+    from coach.detection.detectors import detect_lowest_mpac
+"""
