--- prompts/__init__.py
@@ -1,2 +1,0 @@
-"""Prompt package for conversational agent."""
-
