+++ pyproject.toml
@@ -1,0 +1,22 @@
+[project]
+name = "reframing-retirement"
+version = "0.1.0"
+description = "Autonomy-supportive AI coaching for newly retired adults"
+requires-python = ">=3.11"
+license = {text = "MIT"}
+authors = [
+    {name = "Aidan Halley"}
+]
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+python_files = ["test_*.py"]
+python_functions = ["test_*"]
+
+[tool.ruff]
+line-length = 120
+target-version = "py311"
+
+[tool.ruff.lint]
+select = ["E", "F", "I"]
+ignore = ["E501"]
