+++ scripts/cli.py
